move constants into their own directory
